% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamppPhylip.R
\name{stamppPhylip}
\alias{stamppPhylip}
\title{Export to Phylip Format}
\usage{
stamppPhylip(distance.mat, file = "")
}
\arguments{
\item{distance.mat}{the matrix containing the genetic distances generated from stamppNeisD to be converted into Phylip format}

\item{file}{the file path and name to save the Phylip format matrix as}
}
\description{
Converts the genetic distance matrix generated with stamppNeisD into Phylip format and exports it as a text file
}
\details{
The exported Phylip formated text file can be easily imported into sofware packages such as DARWin (Perrier & Jacquemound-Collet 2006) to be used to generate neighbour joining trees
}
\examples{
# import genotype data and convert to allele frequecies
data(potato.mini, package="StAMPP")
potato.freq <- stamppConvert(potato.mini, "r")
# Calculate genetic distance between populations
potato.D.pop <- stamppNeisD(potato.freq, TRUE, "standard")
# Export the genetic distance matrix in Phylip format
\dontrun{stamppPhylip(potato.D.pop, file="potato_distance.txt")}
}
\references{
Perrier X, Jacquemound-Collet JP (2006) DARWin - Dissimilarity Analysis and Representation for Windows. Agricultural Research for Development
}
\author{
Luke Pembleton <lpembleton at barenbrug.com>
}
